SearchActivity зарефакторил в соответствии с MVVM
