+++ app/src/main/java/com/example/playlistmaker/ui/models/ITunesServerResponseStatus.kt
@@ -1,5 +1,7 @@
 package com.example.playlistmaker.ui.models
 
+import com.example.playlistmaker.domain.search.consumer.ConsumerData
+
 enum class ITunesServerResponseStatus {
-    NOTHING_FOUND, CONNECTION_ERROR, SUCCESS, LOADING
+    NOTHING_FOUND, CONNECTION_ERROR, SUCCESS, LOADING, EMPTY_REQUEST
 }
