+++ app/src/main/java/com/example/playlistmaker/ui/search/activity/TrackAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.playlistmaker.ui.search
+package com.example.playlistmaker.ui.search.activity
 
 import android.util.Log
 import android.view.LayoutInflater
